requirements.txt
streamlit
pandas
pdfplumber
openpyxl
xlrd
unicodedata2
python-dateutil
rapidfuzz
tabula-py
numpy